refactor: DTO 파일 세분화 하여 그룹화
