+++ WeatherForecast/WeatherForecast/Model/DTO/Common/Clouds.swift
@@ -1,0 +1,12 @@
+//
+//  Clouds.swift
+//  WeatherForecast
+//
+//  Created by DONGWOOK SEO on 2023/03/20.
+//
+
+import Foundation
+
+struct Clouds: Decodable {
+    let all: Int
+}
