+++ WeatherForecast/WeatherForecast/Model/DTO/Common/MainData.swift
@@ -1,0 +1,23 @@
+//
+//  MainData.swift
+//  WeatherForecast
+//
+//  Created by DONGWOOK SEO on 2023/03/20.
+//
+
+import Foundation
+
+struct MainData: Decodable {
+    let temp, feelsLike, tempMin, tempMax: Double
+    let pressure, humidity, seaLevel, grndLevel: Int
+
+    enum CodingKeys: String, CodingKey {
+        case temp
+        case feelsLike = "feels_like"
+        case tempMin = "temp_min"
+        case tempMax = "temp_max"
+        case pressure, humidity
+        case seaLevel = "sea_level"
+        case grndLevel = "grnd_level"
+    }
+}
