+++ WeatherForecast/WeatherForecast/Model/DTO/Common/Weather.swift
@@ -1,0 +1,13 @@
+//
+//  Weather.swift
+//  WeatherForecast
+//
+//  Created by DONGWOOK SEO on 2023/03/20.
+//
+
+import Foundation
+
+struct Weather: Decodable {
+    let id: Int
+    let main, description, icon: String
+}
