+++ WeatherForecast/WeatherForecast/Model/DTO/Common/Wind.swift
@@ -1,0 +1,14 @@
+//
+//  Wind.swift
+//  WeatherForecast
+//
+//  Created by DONGWOOK SEO on 2023/03/20.
+//
+
+import Foundation
+
+struct Wind: Decodable {
+    let speed: Double
+    let deg: Int
+    let gust: Double
+}
