+++ WeatherForecast/WeatherForecast/Model/DTO/CurrentWeather/CurrentWeather.swift
