+++ WeatherForecast/WeatherForecast/Model/NetworkError.swift
@@ -10,9 +10,9 @@ import Foundation
 enum NetworkError: LocalizedError {
     case invalidURL
     case invalidData
-    case networkingError
-    case parseError
-    case responseError
+    case networking
+    case parse
+    case response
 
     var errorDescription: String? {
         switch self {
@@ -20,11 +20,11 @@ enum NetworkError: LocalizedError {
             return "URL이 잘못되었습니다."
         case .invalidData:
             return "Data 형식이 잘못되었습니다."
-        case .networkingError:
+        case .networking:
             return "네트워킹에 실패했습니다."
-        case .parseError:
+        case .parse:
             return "데이터를 파싱하는데 실패했습니다."
-        case .responseError:
+        case .response:
             return "응답코드가 정상코드가 아닙니다."
         }
     }
